+++ package.json
@@ -20,6 +20,8 @@
     "react": "^18.2.0",
     "react-datepicker": "^4.24.0",
     "react-dom": "^18.2.0",
+    "react-markdown": "^9.0.1",
+    "react-router-dom": "^6.21.0",
     "react-scripts": "^5.0.1",
     "recharts": "^2.10.3",
     "web-vitals": "^2.1.4"
