+++ src/containers/Blog.js
@@ -1,0 +1,8 @@
+import React from "react";
+import ReactMarkdown from "react-markdown";
+
+const Blog = () => {
+  return <ReactMarkdown>*React-Markdown* is **Awesome**</ReactMarkdown>;
+};
+
+export default Blog;
